fix: refactor users
