+++ src/component/admin/users/UsersAdmin.tsx
@@ -1,12 +1,12 @@
 import { useContext } from 'react';
 import UsersList from './UsersList/UsersList';
 import AdminMenu from '../admin-menu';
-import PageContent from '../../../component/common/PageContent/PageContent';
+import PageContent from '../../common/PageContent/PageContent';
 import AccessContext from '../../../contexts/AccessContext';
-import ConditionallyRender from '../../../component/common/ConditionallyRender';
-import { ADMIN } from '../../../component/providers/AccessProvider/permissions';
+import ConditionallyRender from '../../common/ConditionallyRender';
+import { ADMIN } from '../../providers/AccessProvider/permissions';
 import { Alert } from '@material-ui/lab';
-import HeaderTitle from '../../../component/common/HeaderTitle';
+import HeaderTitle from '../../common/HeaderTitle';
 import { Button } from '@material-ui/core';
 import { useStyles } from './index.styles';
 import { useHistory } from 'react-router-dom';
