--- src/component/admin/users/UsersList/index.js
@@ -1,12 +1,0 @@
-import { connect } from 'react-redux';
-import UsersList from './UsersList';
-
-const mapStateToProps = state => {
-    return {
-        location: state.settings.toJS().location || {},
-    };
-};
-
-const Container = connect(mapStateToProps)(UsersList);
-
-export default Container;
